Moved TalkItem component to its own directory
