+++ src/components/TalkItem/TalkItem.tsx
